Move settings management mutations under new directory
