+++ src/schema/v2/partner/Settings/__tests__/createPartnerLocationMutation.test.js
